requirements_minimal.txt
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.9.2
fastapi-cors==0.0.6
python-multipart
databases==0.8.0
psycopg==3.2.3
psycopg-binary==3.2.3
pgvector==0.2.5
weaviate-client==4.9.3
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.19
langchain-groq==0.2.1
langchain-openai==0.2.8
langgraph==0.2.45
langgraph-checkpoint==2.0.2
langgraph-checkpoint-postgres==2.0.2
langgraph-sdk==0.1.35
sentence-transformers==3.3.0
torch==2.5.1
numpy==1.26.4
pypdf==5.1.0
requests==2.32.3
groq==0.11.0
python-dotenv==1.0.1
Jinja2==3.1.4
pytz==2025.2
asyncpg==0.30.0